% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaBetaPlot.R
\name{alphaBetaPlot}
\alias{alphaBetaPlot}
\title{Makes Alpha ~ Beta Plots for BGC results}
\usage{
alphaBetaPlot(
  outlier.list,
  overlap.zero = TRUE,
  qn.interval = TRUE,
  both.outlier.tests = FALSE,
  saveToFile = NULL,
  plotDIR = "./plots",
  device = "pdf",
  neutral.color = "gray",
  alpha.color = "blue",
  beta.color = "red",
  both.color = "purple",
  margins = c(150, 5.5, 5.5, 5.5),
  margin.units = "points",
  padding = 0.1,
  height = 7,
  width = 7,
  dim.units = "in",
  text.size = 18,
  dpi = 300,
  alpha_limits = NULL,
  beta_limits = NULL
)
}
\arguments{
\item{outlier.list}{List of data.frames returned from get_bgc_outliers().
see ?get_bgc_outliers for more info}

\item{overlap.zero}{Boolean. If TRUE, outliers include when the credible
doesn't contain 0}

\item{qn.interval}{Boolean. If TRUE, outliers = outside quantile interval}

\item{both.outlier.tests}{If TRUE, outliers = both overlap.zero and
qn.interval}

\item{saveToFile}{If specified, saves plots to file. The value
for saveToFile should be the prefix for the filename
you want to save to}

\item{plotDIR}{Directory path. If saveToFile is specified, plots are
saved in this directory}

\item{device}{File format to save plot. Supports ggplot2::ggsave devices}

\item{neutral.color}{Color for non-outlier loci. Default = "gray"}

\item{alpha.color}{Color for alpha outlier loci. Default = "blue"}

\item{beta.color}{Color for beta outlier loci. Default = "red"}

\item{both.color}{Color for loci that are alpha and beta outliers. Default = "purple"}

\item{margins}{Vector of margins for phi plot: c(Top, Right, Bottom, Left)
Top is extended to include the Hybrid Index histogram}

\item{margin.units}{Units for margins parameter. Default = "points"}

\item{padding}{Padding to add to X and Y axes. Default=0.1}

\item{height}{Height for plot. Default = 7}

\item{width}{Width for plot. Default = 7}

\item{dim.units}{Units for height and width. Default = "in"}

\item{text.size}{Size for plot text. Default = 18}

\item{dpi}{DPI for saving plot. Default = 300}

\item{alpha_limits}{Limits (e.g., c(1,1) for plotting alpha. Default = NULL}

\item{beta_limits}{Limits (e.g., c(1,1) for plotting beta. Default = NULL}

\item{line.size}{Size of regression lines in plot. Default = 0.25}
}
\description{
This function generates an Alpha x Beta 2D density plot
with polygon hulls encapsulating alpha and beta outliers.
The plot can be printed to the plot
window in Rstudio (default) or saved to file by specifying a file prefix
with the saveToFile option.
}
\details{
Generates Alpha x Beta Density Plots, Highlighting Outliers.
}
\examples{
alphaBetaPlot(outlier.list = outliers, saveToFile = "pop1_phiPlot",
plotDIR = "./bgc_plots", device = "png")

phiPlot(outlier.list = gene.outliers, ,
line.size = 0.1, qn.interval = FALSE, overlap.zero = FALSE,
both.outlier.tests = TRUE, saveToFile = "mus_phiPlot")

phiPlot(outlier.list = full.outliers,  overlap.zero = FALSE)
}
