% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_bgc_output.R
\name{combine_bgc_output}
\alias{combine_bgc_output}
\title{Aggregate BGC Output}
\usage{
combine_bgc_output(results.dir, prefix, thin = NULL, discard = NULL)
}
\arguments{
\item{results.dir}{Path to a directory containing all the BGC results}

\item{prefix}{Prefix to the input files; e.g., "population1"}

\item{thin}{Thin to every n MCMC samples. E.g. thin=2 cuts samples in half}

\item{discard}{Discard first N samples from each BGC run}
}
\value{
A list of data.frames (each data.frame is a BGC parameter)
}
\description{
This function aggregates multiple BGC runs into one file. Runs are joined
by columns and must all contain the same number of post-burnin MCMC
iterations. It assumes that input files are contained in a single directory
and have specific file suffixes. See the README.md file for details on
file suffixes. Files can have population prefixes.
}
\examples{
aggregate.results <- combine_bgc_output(results.dir = "./results",
                                        prefix = "population1")

aggregate.results <- combine_bgc_output(results.dir = "./results",
                                        prefix = "population2",
                                        thin = 2)

aggregate.results <- combine_bgc_output(results.dir = "./results",
                                        prefix = "pop3",
                                        discard = 2000)
}
