% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_important_rasters.R
\name{extractPointValues}
\alias{extractPointValues}
\title{Function to extract raster values at each sample point for raster stack}
\usage{
extractPointValues(env.list)
}
\arguments{
\item{env.list}{List object generated from prepare_rasters()}
}
\value{
List of data.frames with raster values at each sample locality
for each raster layer, and list of raster names
}
\description{
Function to extract raster values at each sample point for raster stack
}
