% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genind2bgc.R
\name{genind2bgc}
\alias{genind2bgc}
\title{Generate inputs for BGC from genind object. Does not use genotype uncertainty.}
\usage{
genind2bgc(
  gen,
  p1,
  p2,
  admix,
  missingPerLoc = 0.5,
  missingPerInd = 0.5,
  subset = NULL,
  drop = TRUE,
  prefix = "bgc"
)
}
\arguments{
\item{gen}{Genind object containing data for parental and admixed populations}

\item{p1}{Character or vector containing population names for the p1 group}

\item{p2}{Character or vector containing population names for the p2 group}

\item{admix}{Character or vector containing population names for the admixed individuals}

\item{missingPerLoc}{Proportion of missing genotypes allowed to retain a locus}

\item{missingPerInd}{Proportion of missing genotypes allowed to retain an individual}

\item{subset}{An optional parameter defining a number of loci to randomly retain}

\item{prefix}{A string giving the prefix for bgc input file names}
}
\description{
This function parses a genind object to create input data frames for
running genomic cline analysis in BGC. It writes the appropriate output files as:
<prefix>_p0in.txt, <prefix>_p1in.txt, and <prefix>_admixedin.txt
}
\examples{
bgc_input <- genind2bgc(genind, p1=c("PopA", "PopB"),
                                        p2="PopC", admix=c("PopD1", "PopD2"))

bgc_input <- genind2bgc(genind, p1=c("PopA", "PopB"),
                                        p2="PopC", admix=c("PopD1", "PopD2"),
                                        missingPerInd=0.75, missingPerLoc=0.25)

bgc_input <- genind2bgc(genind, p1=c("PopA", "PopB"),
                                        p2="PopC", admix=c("PopD1", "PopD2"),
                                        subset=500)
}
