% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genind2introgress.R
\name{genind2introgress}
\alias{genind2introgress}
\title{Generate inputs for Introgress from genind object}
\usage{
genind2introgress(
  gen,
  p1,
  p2,
  admix,
  missingPerLoc = 0.5,
  missingPerInd = 0.5,
  subset = NULL,
  drop = TRUE,
  prefix = "introgress"
)
}
\arguments{
\item{gen}{Genind object containing data for parental and admixed populations}

\item{p1}{Character or vector containing population names for the p1 group}

\item{p2}{Character or vector containing population names for the p2 group}

\item{admix}{Character or vector containing population names for the admixed individuals}

\item{missingPerLoc}{Proportion of missing genotypes allowed to retain a locus}

\item{missingPerInd}{Proportion of missing genotypes allowed to retain an individual}

\item{subset}{An optional parameter defining a number of loci to randomly retain}

\item{drop}{Boolean value defining whether or not to remove monomorphic loci}

\item{prefix}{Prefix for output Introgress files (default="introgress")}
}
\description{
This function parses a genind object to create input data frames for
running genomic cline analysis in the Introgress R package
}
\examples{
introgress_input <- genind2introgress(genind, p1=c("PopA", "PopB"),
                                        p2="PopC", admix=c("PopD1", "PopD2"))

introgress_input <- genind2introgress(genind, p1=c("PopA", "PopB"),
                                        p2="PopC", admix=c("PopD1", "PopD2"),
                                        missingPerInd=0.75, missingPerLoc=0.25,
                                        drop=TRUE)

introgress_input <- genind2introgress(genind, p1=c("PopA", "PopB"),
                                        p2="PopC", admix=c("PopD1", "PopD2"),
                                        subset=500)
}
