% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bgc_outliers.R
\name{get_bgc_outliers}
\alias{get_bgc_outliers}
\title{Parses BGC output and identifies outlier SNPs}
\usage{
get_bgc_outliers(
  df.list,
  admix.pop,
  popmap,
  qn = 0.975,
  loci.file = NULL,
  save.obj = NULL
)
}
\arguments{
\item{df.list}{List of data.frame objects from combine_bgc_output()}

\item{admix.pop}{String identifying admixed population in population map
file}

\item{popmap}{Population map file consisting of two tab-separated columns:
indID   popID (no header line)}

\item{qn}{Upper quantile interval value for determining outlier SNPs.
Default = 0.975}

\item{loci.file}{Path to two-column tab-separated file containing loci names
(in same order as BGC input). Column 1 should be locus
name, column2 should be the position of the SNP.
header line should be: #CHROM POS}

\item{save.obj}{File path. If supplied, saves outlier info as .RDS object.
The value for save.obj should be the filename for the
RDS object}
}
\value{
List containing 3 data.frame objects. 1) outlier info,
2)popmap info, 3) hybrid index info
}
\description{
Parses BGC output and identifies outlier SNPs
}
\examples{
outliers <- get_bgc_outliers(df.list = aggregated.results,
                             admix.pop = "population1",
                             popmap = "./popmap.txt",
                             qn = 0.975,
                             loci.file = "./bgc_loci.txt",
                             save.obj = "./outlierInfo.rds")
}
