% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_bgc_gff.R
\name{join_bgc_gff}
\alias{join_bgc_gff}
\title{Reads in BGC Outlier Data and Joins with GFF Data}
\usage{
join_bgc_gff(prefix, outlier.list, gff.data, scafInfoDIR = "./scaffold_info")
}
\arguments{
\item{prefix}{Prefix for output file}

\item{outlier.list}{List generated from get_bgc_outliers().
See ?get_bgc_outliers.}

\item{gff.data}{Object output from read_parse_GFF(). See ?read_parse_GFF}

\item{scafInfoDIR}{Directory to save output from this function.}
}
\value{
Data.frame containing outliers genes joined with GFF data
}
\description{
This function joins BGC outlier data with info from a GFF file.
It outputs a joined data.frame of BGC loci that match to GFF genes.
The BGC outlier loci names must be GenBank transcript IDs that can be
related to the transcript IDs in the GFF file The BGC outlier object must
also be generated using upstream methods. See ?get_bgc_outliers for
more info.
Read In BGC Data and Join With GFF Annotations
}
\examples{
outlier.genes <- join_bgc_gff(prefix = "population1",
                              outlier.list = outliers,
                              gff.data = gff.df,
                              scafInfoDIR = "./scaffold_info")
}
