% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseGFF.R
\name{parseGFF}
\alias{parseGFF}
\title{Function to Read and Parse a GFF file}
\usage{
parseGFF(gff.filepath)
}
\arguments{
\item{gff.filepath}{Path to GFF file}
}
\value{
data.frame object with useful GFF info
}
\description{
This function takes a GFF file path as input. It will read it, parse it,
and return a data.frame for use with associate_scaffolds_GFF().
}
\examples{
gff <- parseGFF("./genes.gff")
}
