% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_important_rasters.R
\name{partition_raster_bg}
\alias{partition_raster_bg}
\title{Function to make partitions from background and foreground rasters.}
\usage{
partition_raster_bg(
  env.list,
  number.bg.points = 10000,
  bg.raster = 1,
  plotDIR = "./plots",
  showPLOTS = FALSE,
  agg.factor = 2
)
}
\arguments{
\item{env.list}{Object output from prepare_rasters() function}

\item{number.bg.points}{Number of background points to generate}

\item{bg.raster}{Raster to use for background points}

\item{plotDIR}{Directory to save the plots to}

\item{showPLOTS}{Boolean; Whether to print plots to screen}

\item{agg.factor}{A vector of 1 or 2 numbers for the checkerboard
methods}
}
\value{
Object with background points
}
\description{
This function uses the output from prepare_raster() and makes
background partitions using several methods.
}
\examples{
bg <- partition_raster_bg(env.list = envList,
                          number.bg.points = 10000,
                          bg.raster = 1,
                          plotDIR = "./plots",
                          showPLOTS = TRUE,
                          agg.factor = 2)
}
