% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phiPlot.R
\name{phiPlot}
\alias{phiPlot}
\title{Makes Phi ~ Hybrid Index Plots for Alpha and Beta}
\usage{
phiPlot(
  outlier.list,
  popname = "Population 1",
  overlap.zero = TRUE,
  qn.interval = TRUE,
  both.outlier.tests = FALSE,
  line.size = 0.25,
  saveToFile = NULL,
  plotDIR = "./plots",
  device = "pdf",
  neutral.color = "gray",
  neutral.line.color = "black",
  neutral.line.size = 1,
  alpha.color = "blue",
  beta.color = "red",
  margins = c(150, 5.5, 5.5, 5.5),
  margin.units = "points",
  phi.height = 7,
  phi.width = 7,
  phi.units = "in",
  phi.text.size = 18,
  phi.dpi = 300,
  hist.height = 1.5,
  hist.width = 1,
  hist.x.origin = 0,
  hist.y.origin = 1.05,
  hist.color = "gray57",
  hist.fill = "gray57",
  hist.axis.text.size = 12,
  hist.axis.title.size = 12,
  hist.binwidth = 0.1
)
}
\arguments{
\item{outlier.list}{List of data.frames returned from get_bgc_outliers().
see ?get_bgc_outliers for more info}

\item{popname}{Population name to use as the plot title}

\item{overlap.zero}{Boolean. If TRUE, outliers include when the credible
doesn't contain 0}

\item{qn.interval}{Boolean. If TRUE, outliers = outside quantile interval}

\item{both.outlier.tests}{If TRUE, outliers = both overlap.zero and
qn.interval}

\item{line.size}{Size of regression lines in plot. Default = 0.25}

\item{saveToFile}{If specified, saves plots to file. The value
for saveToFile should be the prefix for the filename
you want to save to}

\item{plotDIR}{Directory path. If saveToFile is specified, plots are
saved in this directory}

\item{device}{File format to save plot. Supports ggplot2::ggsave devices}

\item{neutral.color}{Color for non-outlier loci. Default = "gray"}

\item{neutral.line.color}{Color for dashed linear line showing mean
Default = "black"}

\item{neutral.line.size}{Size for dashed linear line showing mean.
Default = 1.0}

\item{alpha.color}{Color for alpha outlier loci. Default = "blue"}

\item{beta.color}{Color for beta outlier loci. Default = "red"}

\item{margins}{Vector of margins for phi plot: c(Top, Right, Bottom, Left)
Top is extended to include the Hybrid Index histogram}

\item{margin.units}{Units for margins parameter. Default = "points"}

\item{phi.height}{Height for Phi plot. Default = 7}

\item{phi.width}{Width for Phi plot. Default = 7}

\item{phi.units}{Units for phi height and width. Default = "in"}

\item{phi.text.size}{Size for Phi plot text. Default = 18}

\item{phi.dpi}{DPI for saving Phi plot. Default = 300}

\item{hist.height}{Height of Hybrid Index histogram in Phi data coordinates
E.g., If you want the histogram to be equal to half the
height of the phi plot, set hist.height to 0.5}

\item{hist.width}{Height of Hybrid Index histogram in Phi data coordinates
E.g., If you want the histogram to be the same width
as the Phi plot, set this to 1.0}

\item{hist.x.origin}{X Origin (in Phi data coordinates) for Hybrid Index
histogram}

\item{hist.y.origin}{Y origin (in Phi data coordinates) for Hybrid Index
histogram. Set to just above the Phi plot by default}

\item{hist.color}{Color of Hybrid Index histogram}

\item{hist.fill}{Fill for Hybrid Index histogram}

\item{hist.axis.text.size}{Text size for histogram x-axis}

\item{hist.axis.title.size}{Title size for histogram x-axis}

\item{hist.binwidth}{Binwidth for Hybrid Index histogram}
}
\description{
This function generates the Phi ~ Hybrid Index genomic cline plot
shown in Gompert et al. (2012). Outlier genomic clines are colored black,
whereas non-outliers are gray. Phi is the Prob. Ancestry for P1.
1-Phi = P(P0 ancestry). The plot can be printed to the plot
window in Rstudio (default) or saved to file by specifying a file prefix
with the saveToFile option.
}
\details{
Calculates Phi and Plots Phi ~ Hybrid Index, Highlighting Outliers.
}
\examples{
phiPlot(outlier.list = outliers, saveToFile = "pop1_phiPlot",
plotDIR = "./bgc_plots", device = "png")

phiPlot(outlier.list = gene.outliers, popname = "Mus musculus",
line.size = 0.1, qn.interval = FALSE, overlap.zero = FALSE,
both.outlier.tests = TRUE, saveToFile = "mus_phiPlot")

phiPlot(outlier.list = full.outliers, popname = "G. gorilla",
overlap.zero = FALSE)
}
