% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lnl.R
\name{plot_lnl}
\alias{plot_lnl}
\title{Plot LnL Traces to Assess Convergence for BGC Log-likelihoods}
\usage{
plot_lnl(df.list, prefix, thin, plotDIR = "./plots")
}
\arguments{
\item{df.list}{List of data.frames containing aggregated BGC results.
See ?combine_bgc_output}

\item{prefix}{Prefix used for input files}

\item{thin}{Thinning parameter used in BGC}

\item{plotDIR}{Directory to save the plots}
}
\description{
This function makes a log-likelihood trace plot. The thin parameter should
equal the thinning parameter used for BGC. This function assumes that the
BGC runs have been aggregated. See ?combine_bgc_output for help. The plot
should be inspected for convergence before proceeding.
Requires ggplot2 package.
}
\examples{
plot_lnl(df.list = aggregated.results, prefix = "population1",
         thin = 40, plotDIR = "./bgc_plots")
}
