% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traces.R
\name{plot_traces}
\alias{plot_traces}
\title{Plot LnL Traces to Assess Convergence for BGC Log-likelihoods}
\usage{
plot_traces(df.list, prefix, plotDIR = "./plots")
}
\arguments{
\item{df.list}{List of data.frames containing aggregated BGC results.
See ?combine_bgc_output}

\item{prefix}{Prefix used for input files}

\item{plotDIR}{Directory to save the plots}
}
\description{
Wrapper function to make BGC parameter trace plots
(LnL, alpha, beta, qa, qb, hi). It assumes that the BGC runs have been
aggregated. See ?combine_bgc_output. The plot should be inspected for
convergence before proceeding further in the pipeline.
}
\examples{
plot_traces(df.list = aggregated.results, prefix = "population1",
         plotDIR = "./bgc_plots")
}
