% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ENMeval.R
\name{prepare_rasters}
\alias{prepare_rasters}
\title{Script to extract raster data for all sampling localities
and find the most important features.}
\usage{
prepare_rasters(
  raster.dir,
  sample.file,
  header = TRUE,
  bb.buffer = 10,
  plotDIR = "./plots"
)
}
\arguments{
\item{raster.dir}{Directory of rasters to load and crop}

\item{sample.file}{CSV file with sample information (sampleID,popID,lat,lon)}

\item{header}{Boolean; Does sample.file have a header line?}

\item{bb.buffer}{Integer; Buffer around sample bounding box.
bb.buffer = bb.buffer * resolution (in arc-seconds)}

\item{plotDIR}{Directory to save plots to}
}
\value{
List with cropped rasters and other info
}
\description{
This script takes as input a directory of rasters, crops them to the
sampling extent, finds the raster values at each sample locality,
and uses MAXENT and ENMeval to determine the most important raster
layers (i.e. features).
}
\details{
Function to load in and crop the raster layers to the sampling extent

This function takes as input a directory of rasters. Only the desired
rasters should be included in raster.dir. You also will need a
comma-delimited sample.file that has four columns in a specific order:
(sampleIDs,populationIDs,latitude,longitude). You can choose if a header
line is present. The rasters will all be put into a stack and cropped to
the extent of the sample localities + bb.buffer.
}
