% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicClinesXenvironment.R
\name{runIntrogress}
\alias{runIntrogress}
\title{Functions to run INTROGRESS and correlate genomic clines X environment}
\usage{
runIntrogress(
  p1.file,
  p2.file,
  admix.file,
  loci.file,
  clineLabels = c("P1", "Het", "P2"),
  minDelt = 0.8,
  prefix = "population1",
  outputDIR = "./plots",
  showPLOTS = FALSE,
  sep = "\\t",
  pop.id = FALSE,
  ind.id = FALSE,
  fixed = FALSE
)
}
\arguments{
\item{p1.file}{Character string; File path for INTROGRESS p1 input file}

\item{p2.file}{Character string; File path for INTROGRESS p2 input file}

\item{admix.file}{Character string; File path for INTROGRESS admixed file}

\item{loci.file}{Character string; File path for INTROGRESS loci file}

\item{clineLabels}{Character vector of length == 3 for c(P1, Het, P2)
populations}

\item{minDelt}{Numeric; Minimum allele frequency delta to retain loci for
INTROGRESS}

\item{prefix}{Character string; Population prefix for input and output files}

\item{outputDIR}{Character string; File path for INTROGRESS output}

\item{showPLOTS}{Boolean; Whether to print the plots to the screen}

\item{sep}{Character string; Column delimiter for INTROGRESS input files}

\item{pop.id}{Boolean; Value for INTROGRESS parameter pop.id.
See ?prepare.data in introgress R pacakge}

\item{ind.id}{Boolean; Value for INTROGRESS parameter ind.id
See ?prepare.data in introgress R package}

\item{fixed}{Boolean; TRUE if input data are fixed SNPs}
}
\value{
List object containing genomic cline output
}
\description{
These functions run INTROGRESS and generate genomic clines. The genomic
clines are then correlated with the raster values at each sampling locality
for the raster layers determined to be the best predictors for species
distributions in MAXENT. See ?prepare_rasters, ?partition_raster_bg,
?runENMeval, ?summarize_ENMeval, and ?extractPointValues.
}
\details{
Function to run INTROGRESS
}
\examples{
eatt <- runIntrogress(p1.file = "EATT_p1data.txt",
                      p2.file = "EATT_p2data.txt",
                      admix.file = "EATT_admix.txt",
                      loci.file = "EATT_loci.txt",
                      clineLabels = c("EA", "Het", "TT")
                      minDelt = 0.8,
                      prefix = "eatt",
                      outputDIR = "./introgress_plots",
                      sep = "\t",
                      pop.id = FALSE,
                      ind.id = FALSE,
                      fixed = FALSE)
}
