% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicClinesXenvironment.R
\name{subsetIndividuals}
\alias{subsetIndividuals}
\title{Function to subset individuals from the extractRasterPoints data.frame}
\usage{
subsetIndividuals(rasterPoint.df, inds)
}
\arguments{
\item{rasterPoint.df}{data.frame from extractRasterPoints list}

\item{inds}{File path to one-column input file containing subset of
individuals you want to use for clinesXenvironment()}
}
\value{
data.frame with subset of individuals
}
\description{
Function to subset individuals from the extractRasterPoints data.frame
}
