% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_important_rasters.R
\name{runENMeval}
\alias{runENMeval}
\title{Function to run ENMeval and MAXENT on the raster layers.}
\usage{
runENMeval(
  envs.fg,
  bg,
  coords,
  partition.method,
  parallel = FALSE,
  np = 2,
  RMvalues = seq(0.5, 4, 0.5),
  feature.classes = c("L", "LQ", "H", "LQH", "LQHP", "LQHPT"),
  categoricals = NULL,
  agg.factor = 2,
  algorithm = "maxent.jar"
)
}
\arguments{
\item{envs.fg}{First element of envs.list returned from prepare_raster_bg()}

\item{bg}{Object with background points; generated from partition_raster_bg}

\item{coords}{Data.Frame with (lon,lat). 3rd element of envs.list}

\item{partition.method}{Method used for background point partitioning}

\item{parallel}{If TRUE, ENMeval is run parallelized with np CPU cores}

\item{np}{Number of parallel cores to use if parallel = TRUE}

\item{RMvalues}{Vector of non-negative regularization multiplier values.
Higher values impose a stronger penalty on model complexity}

\item{feature.classes}{Character vector of feature classes to be used}

\item{categoricals}{Vector indicating which (if any) of the input
environmental layers are categorical.}

\item{agg.factor}{Aggregation factor(s) for checkerboard
partitioning method.}

\item{algorithm}{Character vector. Defaults to dismo's maxent.jar}
}
\value{
ENMeval object
}
\description{
This function takes as input the object output from preparte_rasters() and
the background partitioning method you would like to use
(see ENMeval documentation). You can visualize how the partitioning
methods will look by viewing the PDF output from partition_raster_bg.
See ?partition_raster_bg for more info.
}
\examples{
eval.par <- runENMeval(envs.fg = envList[[1]],
                       bg = bg,
                       coords = envList[[3]],
                       partition.method = "checkerboard1",
                       parallel = FALSE,
                       RMvalues = seq(0.5, 4, 0.5),
                       feature.classes = c("L",
                                           "LQ",
                                           "H",
                                           "LQH",
                                           "LQHP",
                                           "LQHPT")
                       categoricals = NULL,
                       agg.factor = 2,
                       algorithm = "maxent.jar")
}
